switched to doubles
git-svn-id: https://ws10smt.googlecode.com/svn/trunk@83 ec762483-ff6d-05da-a07a-a48fb63a330f

+++ gi/pyp-topics/src/pyp-topics.hh
@@ -12,7 +12,7 @@ class PYPTopics {
 public:
   typedef std::vector<int> DocumentTopics;
   typedef std::vector<DocumentTopics> CorpusTopics;
-  typedef long double F;
+  typedef double F;
 
 public:
   PYPTopics(int num_topics) : m_num_topics(num_topics), m_word_pyps(1) {}
